+++ com.unity.visualeffectgraph/Editor/Models/Operators/VFXOperatorUtility.cs
@@ -513,7 +513,9 @@ namespace UnityEditor.VFX
         static public VFXExpression FixedRandom(VFXExpression hash, bool perElement)
         {
             VFXExpression seed = new VFXExpressionBitwiseXor(hash, VFXBuiltInExpression.SystemSeed);
-            return new VFXExpressionFixedRandom(seed, perElement);
+            if (perElement)
+                seed = new VFXExpressionBitwiseXor(new VFXAttributeExpression(VFXAttribute.ParticleId), seed);
+            return new VFXExpressionFixedRandom(seed);
         }
 
         public enum SequentialAddressingMode