+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [6.4.0-preview] - 2019-02-21 
 
+- Visual Effect Inspector Cosmetic Improvements
+
 ## [6.3.0-preview] - 2019-02-18
 
 ## [6.2.0-preview] - 2019-02-15